Map the '/' route to an instant redirect. (#65)
Previously we used a nested index route with a <Navigate> element to redirect.
Because HomePage wraps the <Outlet/> element in a <Loading> element which does
not render until the SuiObjects data comes in, this had the effect of redirecting
the '/' to '/tokens' after an indeterminate amount of time. This is unexpected
and undesired, we want '/' to redirect immediately.