Remove (broken) facility for Growthbook to control the default chain.
TL;DR: The old growthbook in feature-gating.ts stopped
working, so all this code was just an obscure way of saying
to default the chain to Testnet. Since we are soon moving
to mainnet and will never again need to control the default
chain from Growthbook, this code just hardcodes it. Currently
to testnet and soon we'll change it to mainnet.